requirements.txt
# Vanilla Views
django-vanilla-views==1.0.4
